Move the websocket shimming code to a separate package
